chore: allow full Docker Model Runner base URL specification
Docker Model Runner will evenutally support multiple local backends
(e.g. vLLM), so we want the user to be able to specify the engine
explicitly. This will also allow specification of the "default" local
engine (i.e. the one that lives at `/engines/v1` with no name
component), which is currently llama.cpp, but may be configurable (via
Docker Desktop) in the future.